# swagger.yaml
openapi: 3.0.0
info:
  title: Mi ecommerce
  version: 1.0.0
  description: |-
    Esta es la API de Mi ecommerce.
    
    Links:
    - [Repositorio de Github](https://github.com/alejandroduranroveta/Sprint2--MyEcommerce)
servers:
  - url: http://localhost:3001/api/v2
    description: Servidor de la api
...
paths:
...
  /products/{id}:
    get:
      tags: 
        - Productos
      summary: Recupera el producto con el id solicitado. Responde con la información completa del producto con el id buscado.
      security:
        - bearerAuth: []          
      parameters:
        - in: path
          name: id
          schema:
            type: number
          required: true
          description: Id del producto a buscar
      responses: 
        '200':
          description: El producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
        '404':
          description: Producto no encontrado
        '500':
          description: Error del servidor
    put:
      tags:
        - Productos
      summary: Actualiza el producto identificado con id. Debe recibir un body con la info del producto a modificar. Responde con la info completa del producto modificado.
      security:
        - bearerAuth: []          
      parameters:
        - in: path
          name: id
          schema:
            type: number
          required: true
          description: Id del producto a editar     
      requestBody:
        description: Editar el producto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Producto'
        required: true
      responses: 
        '200':
          description: El producto fue editado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producto'
        '400':
          description: Bad request
        '404':
          description: Producto no encontrado
        '500':
          description: Error del servidor
    delete:
      tags:
        - Productos
      summary: Eliminar el producto identificado con id. Cuando se elimina un producto, primero debe quitarse de todos los carritos, ademas de las imagenes. 
      security:
        - bearerAuth: []          
      parameters:
        - in: path
          name: id
          schema:
            type: number
          required: true
          description: Id del producto a eliminar
      responses: 
        '200':
          description: El producto fue eliminado
        '400':
          description: Bad request
        '404':
          description: producto no encontrado
        '500':
          description: Error del servidor
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Producto:
      type: object
      properties:
        id:
          type: number
          example: 1
        title: 
          type: string
          example: Manzanas rojas
        price: 
          type: number
          example: 49
        description:
          type: string
          example: Manzanas deliciosas y probablemente naturales
        image:
          type: string
          example: shorturl.at/IJK45
        gallery:
          type: array
          items:
            type: object
            properties:
              picture_id: 
                type: number
                example: 1
              picture_url:
                type: string
                example: shorturl.at/IJK45
        category:
          type: string
          example: Fruta
        mostwanted:
          type: boolean
          example: false
        stock:
          type: number
          example: 0
...